pyproject.toml
[tool.poetry]
name = "peewee-migrate"
version = "1.13.0"
homepage = "https://github.com/klen/peewee_migrate"
repository = "https://github.com/klen/peewee_migrate"
description = "Support for migrations in Peewee ORM"
readme = "README.rst"
authors = ["Kirill Klenov <horneds@gmail.com>"]
license = "MIT"
keywords = ["peewee", "migrations", "orm"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Utilities",
]

[tool.poetry.scripts]
pw_migrate = "peewee_migrate.cli:cli"
pw-migrate = "peewee_migrate.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3"
click = "*"

[tool.poetry.group.dev.dependencies]
black = "*"
pre-commit = "*"
psycopg2-binary = "^2.9.5"
pytest = "*"
pytest-mypy = "*"
ruff = "*"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["ALL"]
ignore = ["ANN", "ARG", "COM", "D", "N", "UP", "S", "SLF", "INP001", "PLR2004"]

[tool.pytest.ini_options]
addopts = "-svx --mypy"

[tool.mypy]
packages = ["peewee_migrate"]
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py310"]
preview = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
